+++ include/pops/radial_kernel.hpp
@@ -22,6 +22,7 @@
 #include "deterministic_kernel.hpp"
 #include "kernel_types.hpp"
 #include "utils.hpp"
+#include "von_mises_distribution.hpp"
 
 #include <cmath>
 #include <map>
@@ -32,61 +33,6 @@
 #include <stdexcept>
 
 namespace pops {
-
-// we need to bring sqrt to the namespace
-// otherwise only candidates are raster-related
-using std::sqrt;
-
-/*! Von Mises Distribution (Circular data distribution)
-
-    mu is the mean angle, expressed in radians between 0 and 2*pi,
-    and kappa is the concentration parameter, which must be greater
-    than or equal to zero. If kappa is equal to zero, this distribution
-    reduces to a uniform random angle over the range 0 to 2*pi.
-*/
-class von_mises_distribution
-{
-public:
-    von_mises_distribution(double mu, double kappa)
-        : mu(mu), kappa(kappa), distribution(0.0, 1.0)
-    {}
-    template<class Generator>
-    double operator()(Generator& generator)
-    {
-        double a, b, c, f, r, theta, u1, u2, u3, z;
-
-        if (kappa <= 1.e-06)
-            return 2 * PI * distribution(generator);
-
-        a = 1.0 + sqrt(1.0 + 4.0 * kappa * kappa);
-        b = (a - sqrt(2.0 * a)) / (2.0 * kappa);
-        r = (1.0 + b * b) / (2.0 * b);
-
-        while (true) {
-            u1 = distribution(generator);
-            z = cos(PI * u1);
-            f = (1.0 + r * z) / (r + z);
-            c = kappa * (r - f);
-            u2 = distribution(generator);
-            if (u2 <= c * (2.0 - c) || u2 < c * exp(1.0 - c))
-                break;
-        }
-
-        u3 = distribution(generator);
-        if (u3 > 0.5) {
-            theta = fmod(mu + acos(f), 2 * PI);
-        }
-        else {
-            theta = fmod(mu - acos(f), 2 * PI);
-        }
-        return theta;
-    }
-
-private:
-    double mu;
-    double kappa;
-    std::uniform_real_distribution<double> distribution;
-};
 
 /*! Get a corresponding enum value for a string which direction.
  *
@@ -150,7 +96,7 @@ protected:
     std::exponential_distribution<double> exponential_distribution;
     bool deterministic_;
     DeterministicDispersalKernel<IntegerRaster> deterministic_kernel;
-    von_mises_distribution von_mises;
+    VonMisesDistribution von_mises;
 
 public:
     RadialDispersalKernel(
